requirements.txt
alembic==1.0.0
asn1crypto==0.24.0
cffi==1.11.5
click==6.7
cryptography==2.3
Flask==1.0.2
Flask-Login==0.4.1
Flask-Migrate==2.2.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-Uploads==0.2.1
Flask-WTF==0.14.2
idna==2.7
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
Pillow==5.2.0
pycparser==2.18
PyMySQL==0.9.2
python-dateutil==2.7.3
python-editor==1.0.3
six==1.11.0
SQLAlchemy==1.2.10
Werkzeug==0.14.1
WTForms==2.2.1
